Moved contents of KubernetesClient.Util into KubernetesClient project
