+++ src/KubernetesClient/Util/Common/CollectionsExtensions.cs
@@ -1,6 +1,7 @@
-// ReSharper disable once CheckNamespace
+using System;
+using System.Collections.Generic;
 
-namespace System.Collections.Generic
+namespace k8s.Util.Common
 {
     internal static class CollectionsExtensions
     {
