+++ src/KubernetesClient/Util/Informer/Cache/DeltaType.cs
