+++ src/KubernetesClient/Util/Informer/Cache/IIndexer.cs
