+++ src/KubernetesClient/Util/Informer/Cache/IStore.cs
