+++ src/KubernetesClient/Util/Informer/Cache/Lister.cs
